refactor directories and files add some code for sessions start create DB module
