+++ server/const.py
@@ -1,0 +1,5 @@
+SERVER_FOLDER = './server'
+STATIC_FOLDER = '/static'
+PLAYER_HAND = 0  # use in PvE
+MODE_PVE = 0
+MODE_PVP = 1
