+++ server/static/js/card_manipulations.js
